Bastion host configuration repositories and access keys
The bastion host configuration is deployed on hosts controlling
administrative access to the nodes of an online managed service.
The inventory variables and playbooks are kept in separate repositories.
The bastion configuration playbooks are public and can be shared between
different domains. The inventories are confidential and private to their
service domain.